Added interupts, although they dont work for now. TODO: fix these mofos

+++ libraries/libc/stddef.h
@@ -11,9 +11,9 @@ typedef long long signed   int int64_t;
 typedef long long unsigned int uint64_t;
 
 typedef char* va_list;
-typedef uint32_t size_t;
-typedef int32_t ssize_t;
-typedef uint32_t uintptr_t;
+typedef uint64_t size_t;
+typedef int64_t ssize_t;
+typedef uint64_t uintptr_t;
 
 #define asm __asm__
 #define nullptr (void*)0
